Remodendo a chamada das consultas para endpoint

+++ uri2609/src/main/java/com/devsuperior/uri2609/projections/CategorySumProjection.java
@@ -3,5 +3,5 @@ package com.devsuperior.uri2609.projections;
 public interface CategorySumProjection {
 
 	String getName();
-	Long getSum();
+	Double getSum();
 }
